LS-8099 (#1138)
Unable to remove entity

+++ modules/returns/src/main/java/com/logistimo/returns/service/ReturnsRepository.java
@@ -98,9 +98,11 @@ public class ReturnsRepository extends Repository {
   }
 
   public void deleteReturnsByCustomer(Long customerId) {
-    returnsItemBatchRepository.deleteByCustomerId(customerId);
-    returnsItemRepository.deleteByCustomerId(customerId);
-    returnsRepository.deleteReturnsByCustomerId(customerId);
+    if(customerId!=null) {
+      returnsItemBatchRepository.deleteByCustomerId(customerId);
+      returnsItemRepository.deleteByCustomerId(customerId);
+      returnsRepository.deleteReturnsByCustomerId(customerId);
+    }
   }
 
   public ReturnsTrackingDetailsVO getReturnTrackingDetails(Long returnId) {
